Large push before merge. Raycasting with entity selection works. FpsCamera is now EditorCamera, the ImGui code is in its own layer instead of Window. Started writing a new model class 'FirstPerson' which will act as a container for components necessary for interactive first person gameplay

+++ LukkelEngine/src/LukkelEngine/Physics/Body/RigidBodyFactory.cpp
@@ -14,8 +14,6 @@ namespace LukkelEngine {
 	{
 		int id = World::s_EntitiesInWorld;
 		RigidBody rigidBody(dimensions, offset, mass, bodyType, friction, restitution, inertia);
-		// rigidBody.m_RigidBody->setUserPointer((void*)id);
-		// rigidBody.setCollisionFlags(btCollisionObject::CF_STATIC_OBJECT);
 		rigidBody.setCollisionFlags(bodyType);
 		return rigidBody;
 	}
@@ -26,7 +24,7 @@ namespace LukkelEngine {
 		RigidBody rigidBody(dimensions, offset, mass, bodyType, friction, restitution, inertia);
 		rigidBody.setID(id);
 		rigidBody.m_RigidBody->setUserIndex(id);
-		rigidBody.m_RigidBody->setUserPointer((void*)(int)id);
+		rigidBody.m_RigidBody->setUserPointer((void*)((uint64_t)id));
 		LKLOG_INFO("RigidBody ID: {0}", id);
 		rigidBody.setCollisionFlags(bodyType);
 		return rigidBody;
